moving packages and creating empty classes
