+++ row-your-boat/src/main/java/com/codeaddi/row_your_boat/controller/sessions/SessionsService.java
@@ -1,0 +1,4 @@
+package com.codeaddi.row_your_boat.controller.sessions;
+
+public class SessionsService {
+}
